requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
msal==1.35.1
requests==2.32.5
pandas==3.0.1
openpyxl==3.1.5
python-multipart==0.0.22
sse-starlette==3.3.3
llama-cpp-python==0.2.90
pydantic==2.12.5
pydantic-settings==2.13.1
numpy==2.4.2
cryptography==46.0.5
pyjwt==2.12.1
tqdm==4.67.3
